requirements.txt
Flask==2.3.2
firebase-admin==6.2.0
Pillow==10.0.0